Refactor build scripts
This makes the logging optional and wraps the build script into yarn commands.
