+++ Procfile
@@ -1,3 +1,3 @@
-webpack-ui: node scripts/webpack-watch
+webpack-ui: node scripts/build -w
 webpack-dev-server: ./scripts/webpack-preview.sh
 rails-server: ./scripts/rails-preview.sh
