+++ package.json
@@ -23,6 +23,7 @@
     "babel-loader": "^8.2.0",
     "blink-diff": "^1.0.13",
     "change-case": "^4.1.1",
+    "commander": "^7.2.0",
     "css-loader": "^5.0.0",
     "cypress": "^6.4.0",
     "cypress-plugin-snapshots": "^1.4.4",
@@ -53,6 +54,10 @@
     "yargs": "^16.2.0"
   },
   "scripts": {
+    "build": "node scripts/build.js",
+    "build:verbose": "node scripts/build.js -v",
+    "watch": "node scripts/build.js -w",
+    "dev": "./cleanstart.sh",
     "format-check": "yarn prettier -c *.js src cypress",
     "format-write": "yarn prettier -w *.js src cypress",
     "lint": "eslint *.js src cypress",
