+++ pre-release.sh
@@ -43,7 +43,7 @@ bundle
 bundle config set --local frozen false
 
 echo "Build library"
-NODE_ENV=production node scripts/webpack-build.js
+NODE_ENV=production node scripts/build.js
 
 echo "Update version.rb file"
 echo -e "module AblyUi\n  VERSION = '$RUBY_VERSION'\nend" > ./lib/ably_ui/version.rb
