--- scripts/webpack-build.js
@@ -1,31 +1,0 @@
-const webpack = require("webpack");
-
-const config =
-  process.env.NODE_ENV === "production"
-    ? require("../webpack.prod")
-    : require("../webpack.dev");
-const setupModules = require("./setup-modules");
-
-const compiler = webpack(config);
-
-compiler.run((err, stats) => {
-  if (err) {
-    console.error(err.stack || err);
-
-    if (err.details) {
-      console.error(err.details);
-    }
-
-    return;
-  }
-
-  console.log(
-    stats.toString({
-      preset: "minimal",
-      chunks: false,
-      colors: true,
-    })
-  );
-
-  setupModules();
-});
